requirements.txt
# Data manipulation and analysis
pandas>=1.3.3,<2.0.0
numpy>=1.21.2,<2.0.0

# Data visualization
matplotlib>=3.4.3,<4.0.0
seaborn>=0.11.2,<1.0.0
plotly>=5.0.0,<6.0.0

# Jupyter environment
jupyter>=1.0.0,<2.0.0

# Machine learning
scikit-learn>=0.24.2,<1.0.0

# Dimensionality reduction and clustering
umap-learn>=0.5.1,<1.0.0

# Statistical analysis
scipy>=1.7.0,<2.0.0

# Development tools
pytest>=7.0.0,<8.0.0
black>=22.0.0,<23.0.0
flake8>=4.0.0,<5.0.0
